Whoaa we got 15%

+++ src/helper.rs
@@ -1,5 +1,5 @@
 #[inline]
-pub fn max_index(v: &Vec<f32>) -> usize {
+pub fn max_index(v: &Vec<f64>) -> usize {
     let mut index = 0;
     let mut max = v[0];
     for i in 1..v.len() {
